api_testing-main/requirements.txt
requests~=2.23.0
pytest~=6.1.2
allpairspy~=2.5.1
PyYAML~=6.0.1